[DLMED] update float64

+++ monai/networks/utils.py
@@ -96,16 +96,16 @@ def normalize_transform(
             corner pixels rather than the image corners.
             See also: https://pytorch.org/docs/stable/nn.functional.html#torch.nn.functional.grid_sample
     """
-    norm = torch.tensor(shape, dtype=torch.float64, device=device)  # no in-place change
+    norm = torch.tensor(shape, dtype=torch.float, device=device)  # no in-place change
     if align_corners:
         norm[norm <= 1.0] = 2.0
         norm = 2.0 / (norm - 1.0)
-        norm = torch.diag(torch.cat((norm, torch.ones((1,), dtype=torch.float64, device=device))))
+        norm = torch.diag(torch.cat((norm, torch.ones((1,), dtype=torch.float, device=device))))
         norm[:-1, -1] = -1.0
     else:
         norm[norm <= 0.0] = 2.0
         norm = 2.0 / norm
-        norm = torch.diag(torch.cat((norm, torch.ones((1,), dtype=torch.float64, device=device))))
+        norm = torch.diag(torch.cat((norm, torch.ones((1,), dtype=torch.float, device=device))))
         norm[:-1, -1] = 1.0 / torch.tensor(shape, dtype=torch.float64, device=device) - 1.0
     norm = norm.unsqueeze(0).to(dtype=dtype)
     norm.requires_grad = False
